temp display fix

+++ src/main.cpp
@@ -247,7 +247,7 @@ void displayScreen(tempSensor t) {
   display.setTextFont(6);
   int sign = 1;
   if (t.temp < 0.0 ) sign = -1;
-  display.printf("%.0f", t.temp);
+  display.printf("%d", (int)t.temp);
   display.setTextFont(4);
   // Set Precision
   double value = (t.temp - int(t.temp)) * sign;
